organize networking files
