+++ Potluck/Potluck/Managers/MHPNetworkManager.swift
@@ -175,38 +175,4 @@ struct MHPNetworkManager {
             }
         })
     }
-    
-//    func sendVerificationEmail(forUser currentUser: User?, completion: @escaping (Result<Bool, Error> ) -> ()) {
-//        let actionCodeSettings =  ActionCodeSettings.init()
-//        actionCodeSettings.handleCodeInApp = true
-//        if let user = currentUser, let email = user.email {
-//            actionCodeSettings.url = URL(string: "https://tza3e.app.goo.gl/emailVerification/?email=\(email)")
-//            actionCodeSettings.setIOSBundleID(Bundle.main.bundleIdentifier!)
-//            user.sendEmailVerification(completion: { (error) in
-//                if error == nil {
-//                    print("verification email sent")
-//                    completion(.success(true))
-//                } else {
-//                    // handle error
-//                    print("Send Verification email error: \(String(describing: error))")
-//                    completion(.failure(error!))
-//                }
-//            })
-//        }
-//    }
-    
-//    func sendResetPasswordEmail(forEmail email: String, completion: @escaping (Result<Bool, Error> ) -> ()) {
-//        let actionCodeSettings =  ActionCodeSettings.init()
-//        actionCodeSettings.handleCodeInApp = true
-//        actionCodeSettings.url = URL(string: "https://tza3e.app.goo.gl/resetPassword/?email=\(email)")
-//        actionCodeSettings.setIOSBundleID(Bundle.main.bundleIdentifier!)
-//        Auth.auth().sendPasswordReset(withEmail: email) { (error) in
-//            if error == nil {
-//                completion(.success(true))
-//            } else {
-//                completion(.failure(error!))
-//                print("Reset password error: \(String(describing: error))")
-//            }
-//        }
-//    }
 }
